refactor component structure, scss files for specific component styling
