+++ src/app/case-studies/page.tsx
@@ -1,4 +1,4 @@
-import Hero from '../components/Hero';
+import Hero from '../components/Hero/Hero';
 
 export default function CaseStudies() {
   return (
