+++ src/app/components/Footer/Footer.tsx
@@ -1,53 +1,53 @@
 'use client';
-import { Header as USWDSHeader, PrimaryNav } from '@trussworks/react-uswds';
+import { PrimaryNav } from '@trussworks/react-uswds';
+import Link from 'next/link';
 import Image from 'next/image';
 import React from 'react';
-import styles from '../styles/Home.module.scss';
+import { NavigationLink } from '../NavigationLink/NavigationLink';
 import classNames from 'classnames';
-import { basePath } from '../utils/constants';
-import { NavigationLink } from './NavigationLink';
+import { basePath } from '../../utils/constants';
+import styles from './Footer.module.scss';
 
-export default function Header() {
+export default function Footer() {
   const [expanded, setExpanded] = React.useState(false);
   const onClick = () => {
     if (window.innerWidth < 1024) setExpanded((prvExpanded) => !prvExpanded);
   };
 
   const testItemsMenu = [
-    <NavigationLink key="one" href="/our-products" onClick={onClick}>
+    <NavigationLink key="one" href="/" onClick={onClick}>
+      Home
+    </NavigationLink>,
+    <NavigationLink key="two" href="/our-products" onClick={onClick}>
       Our products
     </NavigationLink>,
-    <NavigationLink key="two" href="/case-studies" onClick={onClick}>
+    <NavigationLink key="three" href="/case-studies" onClick={onClick}>
       Case studies
     </NavigationLink>,
-    <NavigationLink key="three" href="/engage-with-us" onClick={onClick}>
+    <NavigationLink key="four" href="/engage-with-us" onClick={onClick}>
       Engage with us
     </NavigationLink>,
   ];
 
   return (
     <>
-      <a className="usa-skipnav" href="#main-content">
-        Skip to main content
-      </a>
-
-      <USWDSHeader basic={true} className="bg-background-teal">
+      <footer className="usa-header--basic bg-background-teal">
         <div className="usa-nav-container flex-vertical-center">
           <div className="usa-navbar">
             <div className="usa-logo">
               <em className="usa-logo__text">
-                <a href={`${basePath}/`} title="<Project title>">
+                <a href="http://cdc.gov" title="<Project title>">
                   <span
                     className={classNames('sr-only', styles.navbarLogoText)}
                   >
-                    Data Integration Building Blocks
+                    CDC US center for disease control and prevention
                   </span>
                   <Image
                     width={200}
                     height={40}
                     alt=""
                     className="margin-x-0"
-                    src={`${basePath}/images/dibbs-logo.svg`}
+                    src={`${basePath}/images/CDC.svg`}
                   />
                 </a>
               </em>
@@ -62,7 +62,7 @@ export default function Header() {
             onToggleMobileNav={onClick}
           />
         </div>
-      </USWDSHeader>
+      </footer>
     </>
   );
 }
