+++ src/app/components/Header/Header.module.scss
@@ -1,0 +1,11 @@
+@media (min-width: 1024px) {
+  .navbarLogoText {
+    font-size: 22px;
+  }
+}
+
+@media (max-width: 1024px) {
+  .navbarLogoText {
+    font-size: 13px;
+  }
+}
