+++ src/app/components/MarkdownComponent/MarkdownComponent.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
-import { getMarkdownContent } from '../utils/markdown';
+import { getMarkdownContent } from '../../utils/markdown';
 
 export default async function MarkdownContent(fileName: string) {
   const markdownContent = await getMarkdownContent(fileName);
