+++ src/app/components/NavigationLink/NavigationLink.module.scss
@@ -1,0 +1,11 @@
+@media (min-width: 1024px) {
+  .navbarItemText {
+    color: #f0f0f0;
+  }
+}
+
+@media (max-width: 1024px) {
+  .navbarItemText {
+    color: #565c65;
+  }
+}
