+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
-import Header from './components/Header';
-import Footer from './components/Footer';
-import USABanner from './components/UsaBanner';
+import Header from './components/Header/Header';
+import Footer from './components/Footer/Footer';
+import USABanner from './components/UsaBanner/UsaBanner';
 import { basePath } from './utils/constants';
 import './globals.scss';
 import './custom-styles.css';
