--- src/app/styles/Home.module.scss
@@ -1,286 +1,0 @@
-.productOfferingsLink {
-  width: 300px;
-  background-color: deeppink !important;
-}
-
-.desktopHeroImage2 {
-  margin-bottom: -132px;
-}
-
-.dekstopHeroImage1 {
-  margin-bottom: -2px;
-}
-
-.buildingBlockImage {
-  margin-bottom: -68px;
-}
-
-.buidlingBlockInfoImage {
-  margin-bottom: 0px;
-}
-
-.productOfferingsImage {
-  margin-bottom: -8px;
-}
-
-.sdkInfoMaxWidth {
-  max-width: 100%;
-}
-
-.letterSpacingNeg1 {
-  letter-spacing: -0.2px;
-}
-
-.footerBar {
-  width: 100px;
-  height: 0px;
-  border: 0.5px solid #ffffff;
-}
-
-.getStartedHeader {
-  line-height: 38px;
-}
-
-.buildingBlockItemHeaderText {
-  font-size: 22px;
-}
-
-@media (min-width: 1024px) {
-  .desktopHeroImage3 {
-    margin-top: -289px;
-    margin-bottom: 0px;
-  }
-
-  .backgroundInvitation {
-    background-color: white !important;
-  }
-
-  .navbarItemText {
-    color: #f0f0f0;
-  }
-
-  .navbarItemText:hover {
-    // border-bottom: 2px solid white !important;
-    // padding-bottom: 8px;
-  }
-
-  .headerText,
-  .buildingBlockItemHeader {
-    text-align: left;
-  }
-
-  .partnerItemHeaders {
-    font-family: 'Public Sans', sans-serif;
-    font-weight: 700;
-    font-size: 32px;
-    line-height: 38px;
-  }
-
-  .dynamicHeader {
-    font-size: 40px;
-  }
-
-  .copyrightText {
-    font-size: 14px;
-    text-align: right;
-  }
-
-  .subBodyTextLineHeight {
-    line-height: 25.92px;
-  }
-
-  .homeNavItem {
-    display: none !important;
-  }
-
-  .dynamicShadow {
-    box-shadow: 0 1px 0.25rem 0 rgba(0, 0, 0, 0.1);
-  }
-
-  .navbarLogoText {
-    font-size: 22px;
-  }
-}
-
-@media (max-width: 1024px) {
-  .desktopHeroImage3 {
-    margin-top: 0px;
-  }
-
-  .backgroundInvitation {
-    background-color: #28a0cb;
-  }
-
-  .navbarItemText {
-    color: #565c65;
-  }
-
-  .headerText,
-  .buildingBlockItemHeader {
-    text-align: center;
-  }
-
-  .partnerItemHeaders {
-    font-family: 'Public Sans', sans-serif;
-    font-weight: 400;
-    font-size: 22px;
-    line-height: 26px;
-  }
-
-  .dynamicHeader {
-    font-size: 32px;
-  }
-
-  .copyrightText {
-    font-size: 12px;
-    text-align: center;
-  }
-
-  .subBodyTextLineHeight {
-    line-height: 22.68px;
-  }
-
-  .homeNavItem {
-    display: inline;
-  }
-
-  .dynamicShadow {
-    box-shadow: none;
-  }
-
-  .navbarLogoText {
-    font-size: 13px;
-  }
-}
-
-.main {
-  min-height: 100vh;
-  padding: 4rem 0;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.footer {
-  display: flex;
-  flex: 1;
-  padding: 2rem 0;
-  border-top: 1px solid #eaeaea;
-  justify-content: center;
-  align-items: center;
-}
-
-.footer a {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-grow: 1;
-}
-
-.title a {
-  color: #0070f3;
-  text-decoration: none;
-}
-
-.title a:hover,
-.title a:focus,
-.title a:active {
-  text-decoration: underline;
-}
-
-.title {
-  margin: 0;
-  line-height: 1.15;
-  font-size: 4rem;
-}
-
-.title,
-.description {
-  text-align: center;
-}
-
-.description {
-  margin: 4rem 0;
-  line-height: 1.5;
-  font-size: 1.5rem;
-}
-
-.code {
-  background: #fafafa;
-  border-radius: 5px;
-  padding: 0.75rem;
-  font-size: 1.1rem;
-  font-family:
-    Menlo,
-    Monaco,
-    Lucida Console,
-    Liberation Mono,
-    DejaVu Sans Mono,
-    Bitstream Vera Sans Mono,
-    Courier New,
-    monospace;
-}
-
-.grid {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
-  max-width: 800px;
-}
-
-.card {
-  margin: 1rem;
-  padding: 1.5rem;
-  text-align: left;
-  color: inherit;
-  text-decoration: none;
-  border: 1px solid #eaeaea;
-  border-radius: 10px;
-  transition:
-    color 0.15s ease,
-    border-color 0.15s ease;
-  max-width: 300px;
-}
-
-.card:hover,
-.card:focus,
-.card:active {
-  color: #0070f3;
-  border-color: #0070f3;
-}
-
-.card h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
-}
-
-.card p {
-  margin: 0;
-  font-size: 1.25rem;
-  line-height: 1.5;
-}
-
-.logo {
-  height: 1em;
-  margin-left: 0.5rem;
-}
-
-@media (max-width: 600px) {
-  .grid {
-    width: 100%;
-    flex-direction: column;
-  }
-}
-
-.buttonAccent {
-  background-color: white !important;
-  box-shadow: none !important;
-}
-
-.buttonAccent:hover {
-  background-color: rgb(245, 245, 245) !important;
-  box-shadow: none !important;
-}
